Migrate CCTP message comparison test from Jest to Chai
