+++ tsconfig.eslint.json
@@ -1,6 +1,15 @@
 {
   "extends": "./tsconfig.json",
-  "include": [".eslintrc.js", "index.ts", "./src", "./scripts", "./test", "./deploy", "./tasks"],
+  "include": [
+    ".eslintrc.js",
+    "index.ts",
+    "./src",
+    "./scripts",
+    "./test",
+    "./test/**/*",
+    "./deploy",
+    "./tasks"
+  ],
   "exclude": ["artifacts", "cache", "dist", "node_modules"],
   "compilerOptions": {
     "target": "es6",
