+++ tsconfig.test.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {},
-  "include": ["./test", "./src"]
+  "include": ["./test", "./test/**/*", "./src"]
 }
